moved ErrorStateWithButton out of common
